test(reader): split module switch coverage
Why:
- Continue the test architecture migration by moving pure reader module-switch planner coverage out of the app-host test bundle.
- Keep Android current-document switching contracts covered without paying simulator app-host cost for package-only logic.

What Changed:
- Added ReaderModuleSwitchCoordinatorTests under the BibleUI package test target.
- Removed the migrated module-switch planner assertions from AndBibleTests+AppAndReader.
- Updated the migration ledger for the module-switch coordinator slice.

Validation:
- git diff --check
- python3 scripts/check_repo_standards.py docblocks --all-files
- python3 scripts/check_repo_standards.py source-guards
- xcodebuild test -scheme BibleUITests -only-testing:BibleUITests/ReaderModuleSwitchCoordinatorTests
- xcodebuild test -scheme AndBible -only-testing:AndBibleTests/AndBibleTests/testBibleToolbarMenuRoutesThroughAnchoredQuickSelectorInsteadOfSheet
- GitNexus impact reviewed for documentSwitchPlan, moduleOnlySwitchPlan, and categorySwitchPlan.
- GitNexus detect_changes reported low risk with no affected execution flows.
- Adversarial review completed with no findings.

Impact:
- Reduces app-host unit-test load while preserving reader module-switch planner behavior coverage.
- No production behavior changes.
